修复kline查询接口limit,startTime,endTime参数映射bug

+++ xchange-binance/src/service/market_data_service.rs
@@ -37,10 +37,10 @@ impl BinanceMarketDataService {
             future_exchange_info() -> Result<BinanceExchangeInfo,BinanceError>,
             last_kline(pair: CurrencyPair, interval: KlineInterval) -> Result<BinanceKline, BinanceError>,
             klines_default_limit(pair: CurrencyPair,interval: KlineInterval) -> Result<Vec<BinanceKline>, BinanceError>,
-            klines(pair: CurrencyPair,interval: KlineInterval,limit: Option<u32>,start_time: Option<u64>,end_time: Option<u64>) -> Result<Vec<BinanceKline>, BinanceError>,
+            klines(pair: CurrencyPair,interval: KlineInterval,limit: Option<u16>,start_time: Option<u64>,end_time: Option<u64>) -> Result<Vec<BinanceKline>, BinanceError>,
             future_last_kline(pair: CurrencyPair, interval: KlineInterval) -> Result<BinanceKline, BinanceError>,
             future_klines_default_limit(pair: CurrencyPair,interval: KlineInterval) -> Result<Vec<BinanceKline>, BinanceError>,
-            future_klines(pair: CurrencyPair,interval: KlineInterval,limit: Option<u32>,start_time: Option<u64>,end_time: Option<u64>) -> Result<Vec<BinanceKline>, BinanceError>,
+            future_klines(pair: CurrencyPair,interval: KlineInterval,limit: Option<u16>,start_time: Option<u64>,end_time: Option<u64>) -> Result<Vec<BinanceKline>, BinanceError>,
         }
     }
 }
